+++ Libplanet/Tx/Transaction.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Globalization;
 using System.Linq;
 using System.Security.Cryptography;
 using Bencodex;
@@ -35,8 +34,32 @@ namespace Libplanet.Tx
         private const int BytesCacheThreshold = 50 * 1024;
 
         private TxId? _id;
+        private TxMetadata _metadata;
         private byte[] _signature;
         private byte[] _bytes;
+
+        /// <summary>
+        /// Creates a new <see cref="Transaction{T}"/> instance by copying data from a specified
+        /// transaction <paramref name="metadata"/>.
+        /// </summary>
+        /// <param name="metadata">The transaction metadata that contains data to copy.</param>
+        /// <param name="actions">A list of <see cref="IAction"/>s to include.  This can be empty,
+        /// but cannot be <see langword="null"/>.  This goes to the <see cref="Actions"/> property.
+        /// </param>
+        /// <param name="signature">A digital signature of the content of this
+        /// <see cref="Transaction{T}"/>.  This has to be signed by <paramref name="metadata"/>'s
+        /// <see cref="ITxMetadata.PublicKey"/>. This is copied and then assigned to
+        /// the <see cref="Signature"/> property.</param>
+        /// <exception cref="ArgumentNullException">Thrown when <see langword="null"/>
+        /// is passed to <paramref name="signature"/> or <paramref name="actions"/>.</exception>
+        public Transaction(ITxMetadata metadata, IEnumerable<T> actions, byte[] signature)
+        {
+            _metadata = new TxMetadata(metadata);
+            Actions = actions?.ToImmutableList()
+                ?? throw new ArgumentNullException(nameof(actions));
+            Signature = signature
+                ?? throw new ArgumentNullException(nameof(signature));
+        }
 
         /// <summary>
         /// Creates a new <see cref="Transaction{T}"/>.
@@ -53,15 +76,12 @@ namespace Libplanet.Tx
         /// <see cref="Transaction{T}"/>s committed by the <see cref="Signer"/>
         /// of this transaction.  This goes to the
         /// <see cref="Transaction{T}.Nonce"/> property.</param>
-        /// <param name="signer">An <see cref="Address"/> of the account
-        /// who signs this transaction.  If this is not derived from <paramref
-        /// name="publicKey"/> <see cref="InvalidTxPublicKeyException"/> is
-        /// thrown.  This goes to the <see cref="Signer"/> property.</param>
-        /// <param name="publicKey">A <see cref="PublicKey"/> of the account
-        /// who signs this transaction.  If this does not match to <paramref
-        /// name="signer"/> address <see cref="InvalidTxPublicKeyException"/>
-        /// is thrown.  This cannot be <c>null</c>.  This goes to
-        /// the <see cref="PublicKey"/> property.</param>
+        /// <param name="signer">Ignored.  Left only for backward compatibility.  It will be
+        /// completely gone in the future.  See also <paramref name="publicKey"/> parameter's
+        /// description.</param>
+        /// <param name="publicKey">A <see cref="PublicKey"/> used for signing this transaction.
+        /// This cannot be <see langword="null"/>.  This goes to the <see cref="PublicKey"/>
+        /// property, and <see cref="Signer"/> property is also derived from this value.</param>
         /// <param name="genesisHash">A <see cref="HashDigest{SHA256}"/> value
         /// of the genesis which this <see cref="Transaction{T}"/> is made from.
         /// This can be <c>null</c> iff the transaction is contained
@@ -82,10 +102,8 @@ namespace Libplanet.Tx
         /// or it will throw <see cref="InvalidTxSignatureException"/>.
         /// This is copied and then assigned to the <see cref="Signature"/>
         /// property.</param>
-        /// <exception cref="ArgumentNullException">Thrown when <c>null</c>
-        /// is passed to <paramref name="signature"/>,
-        /// <paramref name="actions"/>, or <paramref name="publicKey"/>.
-        /// </exception>
+        /// <exception cref="ArgumentNullException">Thrown when <see langword="null"/>
+        /// is passed to <paramref name="signature"/> or <paramref name="actions"/>.</exception>
         public Transaction(
             long nonce,
             Address signer,
@@ -96,18 +114,20 @@ namespace Libplanet.Tx
             IEnumerable<T> actions,
             byte[] signature)
         {
-            Nonce = nonce;
-            Signer = signer;
-            GenesisHash = genesisHash;
-            UpdatedAddresses = updatedAddresses ??
-                               throw new ArgumentNullException(nameof(updatedAddresses));
+            _metadata = new TxMetadata(publicKey
+                ?? throw new ArgumentNullException(nameof(publicKey)))
+            {
+                Nonce = nonce,
+                GenesisHash = genesisHash,
+                UpdatedAddresses = updatedAddresses
+                    ?? throw new ArgumentNullException(nameof(updatedAddresses)),
+                Timestamp = timestamp,
+            };
+
             Signature = signature ??
                         throw new ArgumentNullException(nameof(signature));
-            Timestamp = timestamp;
             Actions = actions?.ToImmutableList() ??
                       throw new ArgumentNullException(nameof(actions));
-            PublicKey = publicKey ??
-                        throw new ArgumentNullException(nameof(publicKey));
         }
 
         /// <summary>
@@ -117,30 +137,12 @@ namespace Libplanet.Tx
         /// representation of <see cref="Transaction{T}"/> instance.
         /// </param>
         public Transaction(Bencodex.Types.Dictionary dict)
-            : this(new RawTransaction(dict))
         {
-        }
-
-#pragma warning disable SA1118 // Parameter spans multiple line
-        internal Transaction(RawTransaction rawTx)
-            : this(
-                rawTx.Nonce,
-                new Address(rawTx.Signer),
-                new PublicKey(rawTx.PublicKey.ToArray()),
-                rawTx.GenesisHash != ImmutableArray<byte>.Empty
-                    ? new BlockHash(rawTx.GenesisHash.ToArray())
-                    : (BlockHash?)null,
-                rawTx.UpdatedAddresses.Select(
-                    a => new Address(a)
-                ).ToImmutableHashSet(),
-                DateTimeOffset.ParseExact(
-                    rawTx.Timestamp,
-                    TimestampFormat,
-                    CultureInfo.InvariantCulture).ToUniversalTime(),
-                rawTx.Actions.Select(ToAction).ToImmutableList(),
-                rawTx.Signature.ToArray())
-#pragma warning restore SA1118 // Parameter spans multiple line
-        {
+            _metadata = new TxMetadata(dict);
+            Actions = dict.GetValue<List>(TxMetadata.ActionsKey)
+                .Select(ToAction)
+                .ToImmutableList();
+            _signature = dict.GetValue<Binary>(TxMetadata.SignatureKey).ToByteArray();
         }
 
         private Transaction(
@@ -185,13 +187,13 @@ namespace Libplanet.Tx
         }
 
         /// <inheritdoc cref="ITxMetadata.Nonce"/>
-        public long Nonce { get; }
+        public long Nonce => _metadata.Nonce;
 
         /// <inheritdoc cref="ITxMetadata.Signer"/>
-        public Address Signer { get; }
+        public Address Signer => _metadata.Signer;
 
         /// <inheritdoc cref="ITxMetadata.UpdatedAddresses"/>
-        public IImmutableSet<Address> UpdatedAddresses { get; }
+        public IImmutableSet<Address> UpdatedAddresses => _metadata.UpdatedAddresses;
 
         /// <summary>
         /// A digital signature of the content of this
@@ -225,13 +227,13 @@ namespace Libplanet.Tx
         public IImmutableList<T> Actions { get; }
 
         /// <inheritdoc cref="ITxMetadata.Timestamp"/>
-        public DateTimeOffset Timestamp { get; }
+        public DateTimeOffset Timestamp => _metadata.Timestamp;
 
         /// <inheritdoc cref="ITxMetadata.PublicKey"/>
-        public PublicKey PublicKey { get; }
+        public PublicKey PublicKey => _metadata.PublicKey;
 
         /// <inheritdoc cref="ITxMetadata.GenesisHash"/>
-        public BlockHash? GenesisHash { get; }
+        public BlockHash? GenesisHash => _metadata.GenesisHash;
 
         /// <summary>
         /// Decodes a <see cref="Transaction{T}"/>'s
@@ -245,9 +247,6 @@ namespace Libplanet.Tx
         /// <see cref="Signature"/> is invalid or not signed by
         /// the account who corresponds to <see cref="PublicKey"/>.
         /// </exception>
-        /// <exception cref="InvalidTxPublicKeyException">Thrown when its
-        /// <see cref="Signer"/> is not derived from its
-        /// <see cref="PublicKey"/>.</exception>
         /// <seealso cref="Serialize(bool)"/>
         public static Transaction<T> Deserialize(byte[] bytes, bool validate = true)
         {
@@ -556,7 +555,7 @@ namespace Libplanet.Tx
                 {
                     codec = new Codec();
                     byte[] sigDict =
-                        codec.Encode(Dictionary.Empty.Add(RawTransaction.SignatureKey, _signature));
+                        codec.Encode(Dictionary.Empty.Add(TxMetadata.SignatureKey, _signature));
                     var sigField = new byte[sigDict.Length - 1];
                     Array.Copy(sigDict, 1, sigField, 0, sigField.Length);
                     int sigOffset = _bytes.IndexOf(sigField);
@@ -591,7 +590,10 @@ namespace Libplanet.Tx
         /// <a href="https://bencodex.org/">Bencodex</a>
         /// representation of this <see cref="Transaction{T}"/>.</returns>
         public Bencodex.Types.Dictionary ToBencodex(bool sign) =>
-            ToRawTransaction(sign).ToBencodex();
+            _metadata.ToBencodex(
+                Actions.Select(a => a.PlainValue),
+                sign ? ImmutableArray.Create(_signature) : (ImmutableArray<byte>?)null
+            );
 
         /// <summary>
         /// Validates this <see cref="Transaction{T}"/> and throws an appropriate exception
@@ -601,9 +603,6 @@ namespace Libplanet.Tx
         /// <see cref="Transaction{T}.Signature"/> is invalid or not signed by
         /// the account who corresponds to its <see cref="PublicKey"/>.
         /// </exception>
-        /// <exception cref="InvalidTxPublicKeyException">Thrown when its
-        /// <see cref="Signer"/> is not derived from its
-        /// <see cref="Transaction{T}.PublicKey"/>.</exception>
         public void Validate()
         {
             if (Signature.Length == 0 || !PublicKey.Verify(Serialize(false), Signature))
@@ -612,14 +611,6 @@ namespace Libplanet.Tx
                     $"The signature ({ByteUtil.Hex(Signature)}) is failed " +
                     "to verify.";
                 throw new InvalidTxSignatureException(Id, message);
-            }
-
-            if (!new Address(PublicKey).Equals(Signer))
-            {
-                string message =
-                    $"The public key ({ByteUtil.Hex(PublicKey.Format(true))} " +
-                    $"is not matched to the address ({Signer}).";
-                throw new InvalidTxPublicKeyException(Id, message);
             }
         }
 
@@ -631,29 +622,6 @@ namespace Libplanet.Tx
 
         /// <inheritdoc />
         public override int GetHashCode() => Id.GetHashCode();
-
-        internal RawTransaction ToRawTransaction(bool includeSign)
-        {
-            ImmutableArray<byte> genesisHash =
-                GenesisHash?.ToByteArray().ToImmutableArray() ?? ImmutableArray<byte>.Empty;
-            var rawTx = new RawTransaction(
-                nonce: Nonce,
-                signer: Signer.ByteArray,
-                genesisHash: genesisHash,
-                updatedAddresses: UpdatedAddresses.Select(a =>
-                    a.ByteArray).ToImmutableArray(),
-                publicKey: PublicKey.Format(false).ToImmutableArray(),
-                timestamp: Timestamp.ToString(TimestampFormat, CultureInfo.InvariantCulture),
-                actions: Actions.Select(a => a.PlainValue).ToImmutableArray()
-            );
-
-            if (includeSign)
-            {
-                rawTx = rawTx.AddSignature(Signature);
-            }
-
-            return rawTx;
-        }
 
         private static T ToAction(IValue value)
         {